requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
prisma==0.15.0
python-dotenv==1.1.1
google-generativeai==0.8.5
apscheduler==3.11.0
requests==2.32.5
